requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==2.0.1
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
streamlit==1.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.8.0
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
